--- Project.toml
@@ -1,76 +1,0 @@
-name = "Oceananigans"
-uuid = "9e8cae18-63c1-5223-a75c-80ca9d6e9a09"
-authors = ["Climate Modeling Alliance and contributors"]
-version = "0.76.5"
-
-[deps]
-Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
-CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
-CUDAKernels = "72cfdca4-0801-4ab0-bf6a-d52aa10adc57"
-Crayons = "a8cc5b0e-0ffa-5ad4-8c14-923d3ee1735f"
-CubedSphere = "7445602f-e544-4518-8976-18f8e8ae6cdb"
-Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
-FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
-Glob = "c27321d9-0574-5035-807b-f59d2c89b15c"
-IncompleteLU = "40713840-3770-5561-ab4c-a76e7d0d7895"
-InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
-IterativeSolvers = "42fd0dbc-a981-5370-80f2-aaf504508153"
-JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
-KernelAbstractions = "63c18a36-062a-441e-b654-da1e3ab1ce7c"
-LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
-MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
-NCDatasets = "85f8d34a-cbdd-5861-8df4-14fed0d494ab"
-OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
-OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
-PencilArrays = "0e08944d-e94e-41b1-9406-dcf66b6a9d2e"
-PencilFFTs = "4a48f351-57a6-4416-9ec4-c37015456aae"
-Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
-Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Rotations = "6038ab10-8711-5258-84ad-4b1120ba62dc"
-SeawaterPolynomials = "d496a93d-167e-4197-9f49-d3af4ff8fe40"
-SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
-Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
-StructArrays = "09ab397b-f2b6-538f-b94a-2f83cf4a842a"
-Tullio = "bc48ee85-29a4-5162-ae0b-a64e1601d4bc"
-
-[compat]
-Adapt = "^3"
-CUDA = "3.8, 3.9"
-CUDAKernels = "0.3.3, 0.4"
-Crayons = "^4"
-CubedSphere = "0.1, 0.2"
-DocStringExtensions = "^0.8, 0.9"
-FFTW = "^1"
-Glob = "1.3"
-IncompleteLU = "0.2"
-IterativeSolvers = "0.9"
-JLD2 = "^0.4"
-KernelAbstractions = "0.7, 0.8"
-MPI = "0.16, 0.17, 0.18, 0.19"
-NCDatasets = "^0.10, ^0.11, 0.12"
-OffsetArrays = "^1.4"
-OrderedCollections = "1.1"
-PencilArrays = "0.16, 0.17"
-PencilFFTs = "^0.13.5"
-Rotations = "1.0"
-SeawaterPolynomials = "^0.2.3"
-StructArrays = "0.4, 0.5, 0.6"
-Tullio = "0.2, 0.3"
-julia = "^1.6"
-
-[extras]
-BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
-Coverage = "a2441757-f6aa-5fb2-8edb-039e3f45d037"
-DataDeps = "124859b0-ceae-595e-8997-d05f6a7a8dfe"
-InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-TimerOutputs = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
-TimesDates = "bdfc003b-8df8-5c39-adcd-3a9087f5df4a"
-
-[targets]
-test = ["BenchmarkTools", "Coverage", "DataDeps", "InteractiveUtils", "Plots", "Test", "TimerOutputs", "TimesDates", "SafeTestsets"]
