+++ src/Advection/vector_invariant_advection.jl
@@ -112,6 +112,8 @@ end
 @inline U_dot_∇u(i, j, k, grid::AbstractGrid{FT}, scheme::Nothing, U) where FT = zero(FT)
 @inline U_dot_∇v(i, j, k, grid::AbstractGrid{FT}, scheme::Nothing, U) where FT = zero(FT)
 
+const U1 = UpwindBiased{1}
+
 # For vector Invariant downgrading near the boundaries 
 @inline left_biased_interpolate_xᶠᵃᵃ(i, j, k, grid, ::U1, f::Function, VI::Type{<:SmoothnessStencil}, args...) = @inbounds f(i-1, j, k, grid, args...) 
 @inline left_biased_interpolate_yᵃᶠᵃ(i, j, k, grid, ::U1, f::Function, VI::Type{<:SmoothnessStencil}, args...) = @inbounds f(i, j-1, k, grid, args...)
@@ -120,11 +122,3 @@ end
 @inline right_biased_interpolate_xᶠᵃᵃ(i, j, k, grid, ::U1, f::Function, VI::Type{<:SmoothnessStencil}, args...) = @inbounds f(i, j, k, grid, args...) 
 @inline right_biased_interpolate_yᵃᶠᵃ(i, j, k, grid, ::U1, f::Function, VI::Type{<:SmoothnessStencil}, args...) = @inbounds f(i, j, k, grid, args...)
 @inline right_biased_interpolate_zᵃᵃᶠ(i, j, k, grid, ::U1, f::Function, VI::Type{<:SmoothnessStencil}, args...) = @inbounds f(i, j, k, grid, args...)
-
-@inline left_biased_interpolate_xᶠᵃᵃ(i, j, k, grid, ::U1, f::Function, VI::Type{Nothing}, args...) = @inbounds f(i-1, j, k, grid, args...) 
-@inline left_biased_interpolate_yᵃᶠᵃ(i, j, k, grid, ::U1, f::Function, VI::Type{Nothing}, args...) = @inbounds f(i, j-1, k, grid, args...)
-@inline left_biased_interpolate_zᵃᵃᶠ(i, j, k, grid, ::U1, f::Function, VI::Type{Nothing}, args...) = @inbounds f(i, j, k-1, grid, args...)
-
-@inline right_biased_interpolate_xᶠᵃᵃ(i, j, k, grid, ::U1, f::Function, VI::Type{Nothing}, args...) = @inbounds f(i, j, k, grid, args...) 
-@inline right_biased_interpolate_yᵃᶠᵃ(i, j, k, grid, ::U1, f::Function, VI::Type{Nothing}, args...) = @inbounds f(i, j, k, grid, args...)
-@inline right_biased_interpolate_zᵃᵃᶠ(i, j, k, grid, ::U1, f::Function, VI::Type{Nothing}, args...) = @inbounds f(i, j, k, grid, args...)
